moving contentful dir
